Remove fundraiser sign in JS snippet
This was a bad idea from the beginning so I'm removing it and
reimplementing it in champaign.

The idea behind this code is that once someone has signed an action form
(in this case, a petition form), we get the form data in the petition
form's callback and use that to dispatch an action to the redux store.

The problem with this is that we're trusting that the data we receive is
not corrupt, and we're using an internal API that may change
(dispatching an action with a specific payload). This makes it prone to
break if we introduce changes to the store. It's also not linted, type
checked, or even tested code. It's difficult to make *any* guarantees
about any template script code, so it should be used sparingly.

A better approach is to handle this in Champaign's action form related
code. The redux store can be updated and better checks can be made.